Fixed insufficient fund bug
What happened was that no transferring of the coin could be made since everyone started off with 0. So to fix that, I created a temporary wallet called the genesis, and then the genesis wallet would transfer 100 coins to one user. Then, all coins were drained from the genesis block, thus making it useless.